+++ docs/Changelog.md
@@ -2,3 +2,4 @@
 
 ## 0.6.3 (unreleased)
 * hidden [`--integrate-to-system`](Overview.md#bash-completion) argument
+* interfaces migrated to [`mininterface.interfaces`](Overview.md#all-possible-interfaces) to save around 50 ms starting time due to lazy loading