+++ docs/Changelog.md
@@ -1,6 +1,9 @@
 # Changelog
 
-## 0.7.4
+## 0.7.5
+* UI [configuration](Configuration.md)
+
+## 0.7.4 (2025-01-27)
 * Python 3.13 compatible
 * emits a warning when for config file fields, unknown to the model
 
@@ -13,7 +16,7 @@
 ## 0.7.1 (2024-11-27)
 * GUI scrollbars if window is bigger than the screen
 * [non-interactive][mininterface.Mininterface.__enter__] session support
-* [datetime](Types/#mininterface.types.DatetimeTag) support
+* [datetime](Types.md#mininterface.types.DatetimeTag) support
 * nested generics support (a tuple in a list)
 
 ## 0.7.0 (2024-11-08)